Add product page for brass 60% top-plates.
Change navbar product button/link to a dropdown containing links to both
the G-01 page and the brass plate page. To simplify the jinja2 navbar
macro, change it to use the string "products" as the active variable
when the product dropdown should be displayed as active. This is why
g01.php was changed as well in this commit.